+++ packages/pn-commons/src/components/ApiError/ApiErrorWrapper.tsx
@@ -1,0 +1,38 @@
+import { useErrors } from "../../hooks";
+import ApiError from "./ApiError";
+
+interface ApiErrorWrapperCommonProps {
+  apiId?: string;
+}
+
+interface ApiErrorWrapperProps extends ApiErrorWrapperCommonProps {
+  reloadAction?: () => void;
+  mt?: number;
+  mainText?: string;
+}
+
+interface ApiErrorWrapperGeneralProps extends ApiErrorWrapperCommonProps  {
+  errorComponent: JSX.Element;
+}
+
+export const ApiErrorWrapperGeneral: React.FC<ApiErrorWrapperGeneralProps> = ({ apiId, children, errorComponent }) => {
+  const { hasApiErrors } = useErrors();
+
+  const hasParticularApiErrors = hasApiErrors(apiId);
+
+  return <>
+      { !hasParticularApiErrors && children }
+      { hasParticularApiErrors && errorComponent }
+  </>;
+};
+
+const ApiErrorWrapper: React.FC<ApiErrorWrapperProps> = ({ apiId, children, reloadAction, mt, mainText }) => {
+  return <ApiErrorWrapperGeneral 
+    apiId={apiId} 
+    errorComponent={<ApiError onClick={reloadAction} mt={mt} mainText={mainText} />} 
+  >
+    {children}
+  </ApiErrorWrapperGeneral>
+};
+
+export default ApiErrorWrapper;