+++ packages/pn-commons/src/services/message.service.ts
@@ -12,6 +12,7 @@ interface IAppErrorProps {
       message: string;
     };
   };
+  action?: string;
 }
 
 /**
@@ -44,7 +45,7 @@ const getErrorTitleAndMessage = (
  * @param  {IAppErrorProps} error
  * @returns IAppMessage
  */
-export const createAppError = (error: IAppErrorProps): IAppMessage => {
+export const createAppError = (error: IAppErrorProps, options: { show: boolean } = { show: true }): IAppMessage => {
   const e: IAppMessage = {
     id: _.uniqueId(),
     title: '',
@@ -52,7 +53,11 @@ export const createAppError = (error: IAppErrorProps): IAppMessage => {
     status: 200,
     blocking: false,
     toNotify: true,
+    alreadyShown: !options.show,    // if the error is not to be shown, then we set as already shown just from the start
   };
+  if (error.action) {
+    e.action = error.action;
+  }
   e.status = error.response?.status;
   if (error.response?.customMessage) {
     e.title = error.response.customMessage.title;
@@ -96,6 +101,7 @@ export const createAppMessage = (title: string, message: string, status?: number
     blocking: false,
     toNotify: true,
     status,
+    alreadyShown: false,
   };
   return e;
 };