+++ packages/pn-pa-webapp/src/pages/components/NewNotification/PreliminaryInformations.tsx
@@ -14,18 +14,19 @@ import {
 } from '@mui/material';
 import { PhysicalCommunicationType, CustomDropdown } from '@pagopa-pn/pn-commons';
 
-import { NewNotificationFe, PaymentModel } from '../../../models/NewNotification';
+import { NewNotification, PaymentModel } from '../../../models/NewNotification';
 import { GroupStatus } from '../../../models/user';
 import { useAppDispatch, useAppSelector } from '../../../redux/hooks';
 import { setPreliminaryInformations } from '../../../redux/newNotification/reducers';
 import { getUserGroups } from '../../../redux/newNotification/actions';
+import { PreliminaryInformationsPayload } from '../../../redux/newNotification/types';
 import { RootState } from '../../../redux/store';
 import { trackEventByType } from '../../../utils/mixpanel';
 import { TrackEventType } from '../../../utils/events';
 import NewNotificationCard from './NewNotificationCard';
 
 type Props = {
-  notification: NewNotificationFe;
+  notification: NewNotification;
   onConfirm: () => void;
 };
 
@@ -61,14 +62,13 @@ const PreliminaryInformations = ({ notification, onConfirm }: Props) => {
     /** onSubmit validate */
     onSubmit: (values) => {
       if (formik.isValid) {
-        dispatch(setPreliminaryInformations(values));
+        dispatch(setPreliminaryInformations(values as PreliminaryInformationsPayload));
         onConfirm();
       }
     },
   });
 
   const handleChangeTouched = async (e: ChangeEvent) => {
-    console.log('blah');
     formik.handleChange(e);
     await formik.setFieldTouched(e.target.id, true, false);
   };