+++ packages/pn-personafisica-webapp/src/redux/contact/actions.ts
@@ -1,3 +1,4 @@
+import { performThunkAction } from '@pagopa-pn/pn-commons';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
 import { ContactsApi } from '../../api/contacts/Contacts.api';
@@ -8,17 +9,17 @@ import {
   DigitalAddresses,
   LegalChannelType,
 } from '../../models/contacts';
+import { Party } from '../../models/party';
 import { DeleteDigitalAddressParams, SaveDigitalAddressParams } from './types';
 
+export enum CONTACT_ACTIONS  {
+  GET_DIGITAL_ADDRESSES = 'getDigitalAddresses',
+  GET_ALL_ACTIVATED_PARTIES = 'getAllActivatedParties',
+}
+
 export const getDigitalAddresses = createAsyncThunk<DigitalAddresses, string>(
-  'getDigitalAddresses',
-  async (_params: string, { rejectWithValue }) => {
-    try {
-      return await ContactsApi.getDigitalAddresses();
-    } catch (e) {
-      return rejectWithValue(e);
-    }
-  }
+  CONTACT_ACTIONS.GET_DIGITAL_ADDRESSES,
+  performThunkAction(() => ContactsApi.getDigitalAddresses())
 );
 
 export const createOrUpdateLegalAddress = createAsyncThunk<
@@ -125,6 +126,8 @@ export const disableIOAddress = createAsyncThunk<string, string>(
   }
 );
 
-export const getAllActivatedParties = createAsyncThunk('getAllActivatedParties', async () =>
-  ExternalRegistriesAPI.getAllActivatedParties()
+
+export const getAllActivatedParties = createAsyncThunk<Array<Party>,void>(
+  CONTACT_ACTIONS.GET_ALL_ACTIVATED_PARTIES, 
+  performThunkAction(() => ExternalRegistriesAPI.getAllActivatedParties())
 );