+++ packages/pn-validator/src/types/ValidatorBuilders.ts
@@ -1,0 +1,5 @@
+import { ValidatorBuilder } from "../ValidatorBuilder";
+
+export type ValidatorBuilders<TModel> = {
+    [propertyName in keyof TModel]?: ValidatorBuilder<TModel, TModel[propertyName]>
+};